+++ ReservationApp/src/main/java/com/reservation/ticket/entities/Flight.java
@@ -16,26 +16,26 @@ public class Flight {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 
 	@Column(name = "FLIGHT_NUMBER")
 	private String flightNumber;
 	@Column(name = "OPERATING_AIRLINES")
 	private String operatingAirlines;
-	@Column(name ="DEPARTURE_CITY")
+	@Column(name = "DEPARTURE_CITY")
 	private String departureCity;
-	@Column(name ="ARRIVAL_CITY")
+	@Column(name = "ARRIVAL_CITY")
 	private String arrivalCity;
-	@Column(name="DATE_OF_DEPARTURE")
+	@Column(name = "DATE_OF_DEPARTURE")
 	private Date dateOfDeparture;
-	@Column(name="ESTIMATED_DEPARTURE_TIME")
+	@Column(name = "ESTIMATED_DEPARTURE_TIME")
 	private Timestamp estimatedDepartureTime;
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
