+++ ReservationApp/src/main/java/com/reservation/ticket/repos/FlightRepository.java
@@ -9,8 +9,9 @@ import org.springframework.data.repository.query.Param;
 
 import com.reservation.ticket.entities.Flight;
 
-public interface FlightRepository extends JpaRepository<Flight, Integer>{
+public interface FlightRepository extends JpaRepository<Flight,Long>{
 
 	@Query("from Flight where departureCity=:departureCity and arrivalCity=:arrivalCity and dateOfDeparture=:dateOfDeparture ")
 	List<Flight> findFlights(@Param("departureCity")String from,@Param("arrivalCity") String to, @Param("dateOfDeparture")Date date);
+
 }
